Refactoring image adapters and factories
