+++ src/main/java/com/mealtiger/backend/imageio/adapters/BitmapAdapter.java
@@ -8,6 +8,7 @@ import javax.imageio.ImageIO;
 import javax.imageio.ImageWriteParam;
 import javax.imageio.ImageWriter;
 import javax.imageio.stream.ImageOutputStream;
+import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -22,7 +23,7 @@ public class BitmapAdapter implements ImageAdapter {
 
     private final ImageWriteParam params;
 
-    BitmapAdapter() {
+    public BitmapAdapter() {
         imageWriter = ImageIO.getImageWritersByFormatName("bmp").next();
 
         Configurator configurator = new Configurator();
@@ -41,7 +42,21 @@ public class BitmapAdapter implements ImageAdapter {
         params.setCompressionQuality(((float) compressionQuality)/100);
     }
 
-    public byte[] convert(BufferedImage input) throws IllegalStateException, IllegalArgumentException, IOException {
+    public byte[] convert(BufferedImage image) throws IllegalStateException, IllegalArgumentException, IOException {
+        BufferedImage input;
+        if (image.getColorModel().hasAlpha()) {
+            // Remove alpha channel
+            BufferedImage newImage = new BufferedImage(image.getWidth(), image.getHeight(), BufferedImage.TYPE_INT_RGB);
+            Graphics2D graphics2D = newImage.createGraphics();
+            graphics2D.fillRect(0,0, image.getWidth(), image.getHeight());
+            graphics2D.drawImage(image, 0, 0, null);
+            graphics2D.dispose();
+
+            input = newImage;
+        } else {
+            input = image;
+        }
+
         try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
                 ImageOutputStream imageOutputStream = ImageIO.createImageOutputStream(outputStream)) {
             imageWriter.setOutput(imageOutputStream);
